Move src/utils.ts to src/utils/matchers.ts
